Workshop 10/cine-mate/requirements.txt
streamlit
st-weaviate-connection
weaviate-client
tqdm
toml
pandas
requests
